feat(quartz): 新增定时任务相关功能
- 添加了 AbstractQuartzJob 类作为定时任务的抽象基类
- 新增了多个与定时任务相关的请求和响应对象
- 更新了 LoginController 中的验证码生成逻辑
- 移除了已废弃的 Freemarker 相关代码
- 添加Redis模块，修改表格模块

+++ src/main/java/com/mysiteforme/admin/entity/DTO/TableFieldDTO.java
@@ -12,9 +12,9 @@ public class TableFieldDTO {
 
     private String columnComment;
 
-    private Integer charLength;
+    private Long charLength;
 
-    private Integer numberLength;
+    private Long numberLength;
 
     private Boolean isNullable;
